Split exceptions file
In preparation for PSR-4.
