+++ app/Exceptions/TooSoonForAccessException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Exceptions;
+
+final class TooSoonForAccessException extends LimitedAccessException {
+	public function __construct() {
+		parent::__construct(self::SOON);
+	}
+}
